+++ app/client/event_listeners/listener_authentication.js
@@ -16,7 +16,6 @@ export function listener_authentication(){
 
         if (!authenticationCode)
         {
-            console.log('no code')
             return
         }